Moved Code into a new app Search
-Moved most of the code into apps/search/
