+++ backend/apps/search/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+from .models import Subject, Category
+# Register your models here.
+
+admin.site.register(Subject)
+admin.site.register(Category)
