+++ backend/apps/search/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from .views import SearchView
+
+urlpatterns = [
+  path('', SearchView.as_view(), name='search'),
+]
