+++ backend/apps/users/serializers.py
@@ -1,6 +1,5 @@
 from rest_framework import serializers
 from django.contrib.auth.models import User
-from .models import TutorProfile
 
 class UserSerializer(serializers.ModelSerializer):
   class Meta:
@@ -17,9 +16,3 @@ class UserSerializer(serializers.ModelSerializer):
     instance.save()
     return instance
 
-class TutorProfileSerializer(serializers.ModelSerializer):
-    first_name = serializers.CharField(source="profile.user.first_name")
-    last_name = serializers.CharField(source="profile.user.last_name")
-    class Meta:
-        model = TutorProfile
-        fields = ["first_name", "last_name", "bio", "hourly_rate", "state", "city"]
