+++ backend/backend/urls.py
@@ -21,4 +21,5 @@ urlpatterns = [
   path('admin/', admin.site.urls),
   path('users/', include('apps.users.urls')),
   path('uploads/', include('apps.uploads.urls')),
+  path('search/', include('apps.search.urls')),
 ]
